revert(signals): remove automatic report canvas generation
Reverts the report → canvas auto-generation pipeline (#82144, #82145, #83972, #85108): the SignalReportCanvas/SignalReportCanvasGeneration models, the Temporal generation workflow, the canvas report_canvas API, the backfill command, and the report_canvas OAuth scope preset.

Model removal is state-only (SeparateDatabaseAndState); historical migrations stay in place and the tables are dropped in a follow-up after a deploy cycle.

Generated-By: PostHog Desktop
Task-Id: 8a3abc83-4057-41ec-9344-80a320e65b25
