+++ posthog/temporal/tests/ai/test_module_integrity.py
@@ -103,7 +103,6 @@ class TestSignalsProductModuleIntegrity:
             "SignalsScoutCoordinatorWorkflow",
             "CustomSignalAgentWorkflow",
             "SignalReportInboxNotificationWorkflow",
-            "SignalReportCanvasWorkflow",
         ]
         actual_workflow_names = [w.__name__ for w in SIGNALS_PRODUCT_WORKFLOWS]
         assert len(actual_workflow_names) == len(expected_workflows), (
@@ -143,10 +142,6 @@ class TestSignalsProductModuleIntegrity:
             "mark_report_in_progress_activity",
             "mark_report_pending_input_activity",
             "mark_report_ready_activity",
-            "fail_report_canvas_generation_activity",
-            "report_canvases_enabled_activity",
-            "start_report_canvas_generation_activity",
-            "poll_report_canvas_generation_activity",
             "publish_report_completed_activity",
             "report_has_assigned_signals_activity",
             "revert_report_to_candidate_activity",
