+++ posthog/test/repo_invariants/setup_receivers_baseline.txt
@@ -203,8 +203,6 @@ post_save:products.review_hog.backend.receivers.handle_task_run_saved
 post_save:products.signals.backend.receivers.capture_status_change_analytics
 post_save:products.signals.backend.receivers.close_pr_when_report_dismissed
 post_save:products.signals.backend.receivers.emit_report_embedding_on_document_change
-post_save:products.signals.backend.receivers.mark_report_canvas_collaborative_from_message
-post_save:products.signals.backend.receivers.mark_report_canvas_collaborative_from_version
 post_save:products.signals.backend.receivers.reconcile_report_embedding_on_verdict_saved
 post_save:products.signals.backend.receivers.tombstone_report_embedding_on_delete
 post_save:products.slack_app.backend.signals.invalidate_repo_list_on_user_github_change
