+++ products/canvas/backend/presentation/serializers.py
@@ -128,7 +128,6 @@ class CanvasSerializer(serializers.ModelSerializer):
             "template_id",
             "context",
             "generation_task_id",
-            "discussion_task_id",
             "pinned",
             "pinned_at",
             "current_version_id",
