+++ products/tasks/backend/exceptions.py
@@ -143,13 +143,6 @@ class SandboxMissingRepositoryError(ProcessTaskFatalError):
     pass
 
 
-class RequiredMcpUnavailableError(ProcessTaskFatalError):
-    """A task that cannot run without PostHog MCP cannot reach its configured server."""
-
-    def __init__(self, message: str, context: dict[str, Any]):
-        ProcessTaskError.__init__(self, message, context, None, capture=False, non_retryable=True)
-
-
 class SandboxNotRunningError(SandboxExecutionError):
     """Sandbox is not in a running state."""
