+++ products/tasks/backend/facade/contracts.py
@@ -295,8 +295,6 @@ class TaskActivityDTO:
     latest_comment_id: UUID | None = None
     latest_comment_scope: str | None = None
     latest_comment_item_id: str | None = None
-    target_scope: str | None = None
-    target_id: str | None = None
     is_unread: bool = True
