+++ products/tasks/backend/webhooks.py
@@ -267,12 +267,6 @@ def handle_pull_request_event(payload: dict) -> HttpResponse:
     )
     if task_run is not None and is_internal_branch:
         _record_run_pr_url(task_run, pr_url)
-        # Fired regardless of whether this webhook was the first to record output.pr_url. The agent
-        # server usually records the URL first, so _record_run_pr_url takes its "already recorded"
-        # early return; a canvas the summary workflow built before the PR existed would otherwise
-        # keep implementation_pr_url null forever. The refresh is idempotent: an unchanged report
-        # fingerprint skips generation.
-        _enqueue_report_canvas_refresh(task_run)
 
     # After the backstop on purpose: a just-backfilled pr_url means the run now
     # claims this PR. Gated on the run's *primary* PR — output.pr_state describes
@@ -360,21 +354,6 @@ def handle_pull_request_review_event(payload: dict) -> HttpResponse:
         run_id=str(task_run.id) if task_run else None,
     )
     return HttpResponse(status=200)
-
-
-def _enqueue_report_canvas_refresh(task_run: TaskRun) -> None:
-    """Rebuild the report canvas for this run's task after a PR webhook.
-
-    Best-effort: a Signals import or broker hiccup must never fail the webhook.
-    """
-    try:
-        from products.signals.backend.tasks import (  # noqa: PLC0415 — keeps Signals workers off webhook startup
-            refresh_report_canvases_for_task,
-        )
-
-        refresh_report_canvases_for_task.delay(str(task_run.task_id))
-    except Exception:
-        logger.warning("github_pr_webhook_report_canvas_refresh_failed", task_id=str(task_run.task_id), exc_info=True)
 
 
 def _record_run_pr_url(task_run: TaskRun, pr_url: str) -> None:
